# output/flask/openapi.json
{
  "paths": {
    "/": {
      "get": {
        "operationId": "home",
        "description": "Endpoint home",
        "responses": {
          "200": {
            "description": "Successful response"
          }
        }
      }
    },
    "/api/users": {
      "get": {
        "operationId": "get_users",
        "description": "Endpoint get_users",
        "responses": {
          "200": {
            "description": "Successful response"
          }
        }
      },
      "post": {
        "operationId": "create_user",
        "description": "Endpoint create_user",
        "responses": {
          "200": {
            "description": "Successful response"
          }
        }
      }
    },
    "/api/users/{user_id}": {
      "get": {
        "operationId": "get_user",
        "description": "Endpoint get_user",
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "put": {
        "operationId": "update_user",
        "description": "Endpoint update_user",
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "delete": {
        "operationId": "delete_user",
        "description": "Endpoint delete_user",
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    }
  },
  "info": {
    "title": "main_",
    "version": "1.0.0"
  },
  "openapi": "3.0.2",
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local development server"
    }
  ]
}